systemd: Add file existence dependency on /proc/device-tree/model
Since this deamon requires to open /proc/device-tree/model to get the
board model information, the service sometimes failed to start on the
system which does not use device-tree. For example, the system on the
x86-64 architecture does not use device-tree so that this daemon would
fail to be activated on this system. In order to solve this issue, this
patch adds a ConditionPathExists dependency on /proc/device-tree/model
to the systemd unit configuration file.

Change-Id: Ibf1bdfbe3df76a878ff608464907c84351fe35d1
Signed-off-by: Wook Song <wook16.song@samsung.com>